api/requirements.txt
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.18.0
pycodestyle==2.7.0
toml==0.10.2
Werkzeug==2.0.1

